requirements.txt
pyperclip >= 1.8.2
python-dotenv >= 0.20.0
